requirements.txt

streamlit
streamlit-webrtc
ultralytics
paddleocr
paddlepaddle
opencv-python-headless
numpy
Pillow
torch
torchvision
av